Move virtual table helpers from builder pkg to literals pkg
